+++ Modules/Intent.Modules.ModuleBuilder/Templates/Api/ApiElementExtensionModel/ApiElementExtensionModelTemplatePartial.cs
@@ -1,14 +1,13 @@
 using System.Collections.Generic;
 using Intent.Engine;
 using Intent.Metadata.Models;
-using Intent.Modules.Common.CSharp.Templates;
-using Intent.Modules.Common.Templates;
 using Intent.ModuleBuilder.Api;
 using Intent.Modules.Common;
 using Intent.Modules.Common.CSharp;
+using Intent.Modules.Common.CSharp.Templates;
+using Intent.Modules.Common.Templates;
 using Intent.Modules.Common.Types.Api;
 using Intent.RoslynWeaver.Attributes;
-using Intent.Templates;
 
 [assembly: DefaultIntentManaged(Mode.Merge)]
 [assembly: IntentTemplate("Intent.ModuleBuilder.CSharp.Templates.CSharpTemplatePartial", Version = "1.0")]
@@ -24,6 +23,8 @@ namespace Intent.Modules.ModuleBuilder.Templates.Api.ApiElementExtensionModel
         [IntentManaged(Mode.Merge, Signature = Mode.Fully)]
         public ApiElementExtensionModelTemplate(IOutputTarget outputTarget, ElementExtensionModel model) : base(TemplateId, outputTarget, model)
         {
+            AddNugetDependency(IntentNugetPackages.IntentModulesCommon);
+
             if (Model.TypeReference.Element.Id == FolderModel.SpecializationTypeId)
             {
                 AddNugetDependency(IntentNugetPackages.IntentModulesCommonTypes);
@@ -33,12 +34,15 @@ namespace Intent.Modules.ModuleBuilder.Templates.Api.ApiElementExtensionModel
         public override void BeforeTemplateExecution()
         {
             base.BeforeTemplateExecution();
-            if (Model.TypeReference.Element.Id == FolderModel.SpecializationTypeId && Model.MenuOptions != null)
+
+            if (Model.TypeReference.Element.Id != FolderModel.SpecializationTypeId || Model.MenuOptions == null)
             {
-                foreach (var creationOption in Model.MenuOptions.ElementCreations)
-                {
-                    ExecutionContext.EventDispatcher.Publish(new NotifyModelHasParentFolderEvent(creationOption.TypeReference.Element.Id));
-                }
+                return;
+            }
+
+            foreach (var creationOption in Model.MenuOptions.ElementCreations)
+            {
+                ExecutionContext.EventDispatcher.Publish(new NotifyModelHasParentFolderEvent(creationOption.TypeReference.Element.Id));
             }
         }
 
@@ -54,12 +58,12 @@ namespace Intent.Modules.ModuleBuilder.Templates.Api.ApiElementExtensionModel
             return new ElementSettingsModel((IElement)Model.TypeReference.Element);
         }
 
-        private string FormatForCollection(string name, bool asCollection)
+        private static string FormatForCollection(string name, bool asCollection)
         {
             return asCollection ? $"IList<{name}>" : name;
         }
 
-        private string GetCreationOptionName(ElementCreationOptionModel option)
+        private static string GetCreationOptionName(ElementCreationOptionModel option)
         {
             if (option.GetOptionSettings().ApiModelName() != null)
             {
