+++ Modules/Intent.Modules.ModuleBuilder/Templates/Api/ApiElementModel/ApiElementModelTemplatePartial.cs
@@ -1,13 +1,11 @@
+using System.Collections.Generic;
 using System.Linq;
 using Intent.Engine;
-using Intent.Modules.Common.Templates;
 using Intent.ModuleBuilder.Api;
-using Intent.RoslynWeaver.Attributes;
-using Intent.Templates;
-using System.Collections.Generic;
-using Intent.Modules.Common;
 using Intent.Modules.Common.CSharp;
 using Intent.Modules.Common.CSharp.Templates;
+using Intent.Modules.Common.Templates;
+using Intent.RoslynWeaver.Attributes;
 
 [assembly: DefaultIntentManaged(Mode.Merge)]
 [assembly: IntentTemplate("Intent.ModuleBuilder.CSharp.Templates.CSharpTemplatePartial", Version = "1.0")]
@@ -17,17 +15,16 @@ namespace Intent.Modules.ModuleBuilder.Templates.Api.ApiElementModel
     [IntentManaged(Mode.Merge, Signature = Mode.Fully)]
     partial class ApiElementModelTemplate : CSharpTemplateBase<ElementSettingsModel>
     {
-        public bool HasParentFolder { get; private set; } = false;
-        public List<AssociationSettingsModel> AssociationSettings { get; }
-
         [IntentManaged(Mode.Fully)]
         public const string TemplateId = "Intent.ModuleBuilder.Templates.Api.ApiElementModel";
 
         [IntentManaged(Mode.Ignore)]
         public ApiElementModelTemplate(IOutputTarget outputTarget, ElementSettingsModel model, List<AssociationSettingsModel> associationSettings) : base(TemplateId, outputTarget, model)
         {
+            AddNugetDependency(IntentNugetPackages.IntentModulesCommon);
+
             AssociationSettings = associationSettings;
-            AddTypeSource(CSharpTypeSource.Create(ExecutionContext, ApiElementModelTemplate.TemplateId, collectionFormat: "IEnumerable<{0}>"));
+            AddTypeSource(CSharpTypeSource.Create(ExecutionContext, TemplateId, collectionFormat: "IEnumerable<{0}>"));
             ExecutionContext.EventDispatcher.Subscribe<NotifyModelHasParentFolderEvent>(@event =>
             {
                 if (@event.ModelId == model.Id)
@@ -36,6 +33,10 @@ namespace Intent.Modules.ModuleBuilder.Templates.Api.ApiElementModel
                 }
             });
         }
+
+        public bool HasParentFolder { get; private set; }
+
+        public List<AssociationSettingsModel> AssociationSettings { get; }
 
         protected override CSharpFileConfig DefineFileConfig()
         {
@@ -65,18 +66,12 @@ namespace Intent.Modules.ModuleBuilder.Templates.Api.ApiElementModel
             return string.Join(", ", interfaces);
         }
 
-        private string FormatForCollection(string name, bool asCollection)
+        private static string FormatForCollection(string name, bool asCollection)
         {
             return asCollection ? $"IList<{name}>" : name;
         }
 
-        private string FormatForCollection(AssociationCreationOptionModel option)
-        {
-            var asCollection = option.GetOptionSettings().AllowMultiple();
-            return asCollection ? $"IList<{option.Type.ApiModelName}>" : option.Type.ApiModelName;
-        }
-
-        private string GetCreationOptionName(ElementCreationOptionModel option)
+        private static string GetCreationOptionName(ElementCreationOptionModel option)
         {
             if (option.GetOptionSettings().ApiModelName() != null)
             {
@@ -90,18 +85,6 @@ namespace Intent.Modules.ModuleBuilder.Templates.Api.ApiElementModel
         {
             return Model.GetInheritedType()?.MenuOptions?.ElementCreations.Any(x => x.Type.Id == creationOption.Type.Id) ??
                    false;
-        }
-
-        private bool ExistsInBase(AssociationCreationOptionModel creationOption)
-        {
-            return Model.GetInheritedType()?.MenuOptions?.AssociationCreations.Any(x => x.Type.Id == creationOption.Type.Id) ??
-                   false;
-        }
-
-        private bool ExistsInBase(AssociationSettingsModel associationSettings)
-        {
-            return false;
-            //throw new System.NotImplementedException();
         }
     }
 }
