+++ Modules/Intent.Modules.ModuleBuilder/Templates/Registration/SingleFileListModel/SingleFileListModelTemplateRegistrationTemplatePartial.cs
@@ -1,15 +1,13 @@
 using System.Collections.Generic;
 using System.Linq;
 using Intent.Engine;
-using Intent.Metadata.Models;
+using Intent.ModuleBuilder.Api;
 using Intent.Modules.Common;
 using Intent.Modules.Common.CSharp;
 using Intent.Modules.Common.CSharp.Templates;
 using Intent.Modules.Common.Templates;
 using Intent.Modules.Common.Types.Api;
 using Intent.Modules.Common.VisualStudio;
-using Intent.ModuleBuilder.Api;
-using Intent.Templates;
 
 namespace Intent.Modules.ModuleBuilder.Templates.Registration.SingleFileListModel
 {
@@ -17,8 +15,10 @@ namespace Intent.Modules.ModuleBuilder.Templates.Registration.SingleFileListMode
     {
         public const string TemplateId = "Intent.ModuleBuilder.TemplateRegistration.SingleFileListModel";
 
-        public SingleFileListModelTemplateRegistrationTemplate(IProject project, TemplateRegistrationModel model) : base(TemplateId, project, model)
+        public SingleFileListModelTemplateRegistrationTemplate(IOutputTarget outputTarget, TemplateRegistrationModel model) : base(TemplateId, outputTarget, model)
         {
+            AddNugetDependency(IntentNugetPackages.IntentModulesCommon);
+
             if (!string.IsNullOrWhiteSpace(Model.GetModelType()?.ParentModule.NuGetPackageId))
             {
                 AddNugetDependency(new NugetPackageInfo(Model.GetModelType()?.ParentModule.NuGetPackageId, Model.GetModelType()?.ParentModule.NuGetPackageVersion));
@@ -30,8 +30,11 @@ namespace Intent.Modules.ModuleBuilder.Templates.Registration.SingleFileListMode
         }
 
         public string TemplateName => $"{Model.Name.ToCSharpIdentifier().RemoveSuffix("Template")}Template";
+
         public IList<string> OutputFolders => Model.GetParentFolders().Select(x => x.Name).Concat(new[] { Model.Name }).ToList();
+
         public string FolderPath => string.Join("/", OutputFolders);
+
         public string FolderNamespace => string.Join(".", OutputFolders);
 
         protected override CSharpFileConfig DefineFileConfig()
@@ -42,16 +45,6 @@ namespace Intent.Modules.ModuleBuilder.Templates.Registration.SingleFileListMode
                 relativeLocation: $"{FolderPath}");
         }
 
-        public override IEnumerable<INugetPackageInfo> GetNugetDependencies()
-        {
-            return new INugetPackageInfo[]
-            {
-                IntentNugetPackages.IntentModulesCommon
-            }
-            .Union(base.GetNugetDependencies())
-            .ToArray();
-        }
-        
         private string GetTemplateNameForTemplateId()
         {
             return TemplateName;
