+++ Modules/Intent.Modules.ModuleBuilder/Templates/Registration/SingleFileNoModel/SingleFileNoModelTemplateRegistrationTemplatePartial.cs
@@ -1,15 +1,11 @@
 using System.Collections.Generic;
 using System.Linq;
 using Intent.Engine;
-using Intent.Metadata.Models;
-using Intent.Modules.Common;
+using Intent.ModuleBuilder.Api;
 using Intent.Modules.Common.CSharp;
 using Intent.Modules.Common.CSharp.Templates;
 using Intent.Modules.Common.Templates;
 using Intent.Modules.Common.Types.Api;
-using Intent.Modules.Common.VisualStudio;
-using Intent.ModuleBuilder.Api;
-using Intent.Templates;
 
 namespace Intent.Modules.ModuleBuilder.Templates.Registration.SingleFileNoModel
 {
@@ -17,13 +13,17 @@ namespace Intent.Modules.ModuleBuilder.Templates.Registration.SingleFileNoModel
     {
         public const string TemplateId = "Intent.ModuleBuilder.TemplateRegistration.SingleFileNoModel";
 
-        public SingleFileNoModelTemplateRegistrationTemplate(IProject project, TemplateRegistrationModel model) : base(TemplateId, project, model)
+        public SingleFileNoModelTemplateRegistrationTemplate(IOutputTarget outputTarget, TemplateRegistrationModel model) : base(TemplateId, outputTarget, model)
         {
+            AddNugetDependency(IntentNugetPackages.IntentModulesCommon);
         }
 
         public string TemplateName => $"{Model.Name.ToCSharpIdentifier().RemoveSuffix("Template")}Template";
+
         public IList<string> OutputFolders => Model.GetParentFolders().Select(x => x.Name).Concat(new[] { Model.Name }).ToList();
+
         public string FolderPath => string.Join("/", OutputFolders);
+
         public string FolderNamespace => string.Join(".", OutputFolders);
 
         protected override CSharpFileConfig DefineFileConfig()
@@ -32,16 +32,6 @@ namespace Intent.Modules.ModuleBuilder.Templates.Registration.SingleFileNoModel
                 className: $"{TemplateName}Registration",
                 @namespace: $"{OutputTarget.GetNamespace()}.{FolderNamespace}",
                 relativeLocation: $"{FolderPath}");
-        }
-
-        public override IEnumerable<INugetPackageInfo> GetNugetDependencies()
-        {
-            return new INugetPackageInfo[]
-            {
-                IntentNugetPackages.IntentModulesCommon
-            }
-            .Union(base.GetNugetDependencies())
-            .ToArray();
         }
 
         private string GetTemplateNameForTemplateId()
